+++ src/plugins/workspace/public/application.tsx
@@ -10,6 +10,7 @@ import { OpenSearchDashboardsContextProvider } from '../../opensearch_dashboards
 import { WorkspaceFatalError } from './components/workspace_fatal_error';
 import { WorkspaceCreatorApp } from './components/workspace_creator_app';
 import { WorkspaceUpdaterApp } from './components/workspace_updater_app';
+import { WorkspaceListApp } from './components/workspace_list_app';
 import { Services } from './types';
 
 export const renderCreatorApp = ({ element }: AppMountParameters, services: Services) => {
@@ -44,6 +45,18 @@ export const renderFatalErrorApp = (params: AppMountParameters, services: Servic
   ReactDOM.render(
     <OpenSearchDashboardsContextProvider services={services}>
       <WorkspaceFatalError error={history.location.state?.error} />
+    </OpenSearchDashboardsContextProvider>,
+    element
+  );
+
+  return () => {
+    ReactDOM.unmountComponentAtNode(element);
+  };
+};
+export const renderListApp = ({ element }: AppMountParameters, services: Services) => {
+  ReactDOM.render(
+    <OpenSearchDashboardsContextProvider services={services}>
+      <WorkspaceListApp />
     </OpenSearchDashboardsContextProvider>,
     element
   );