+++ src/plugins/workspace/public/plugin.ts
@@ -18,6 +18,7 @@ import {
   WORKSPACE_OVERVIEW_APP_ID,
   WORKSPACE_CREATE_APP_ID,
   WORKSPACE_UPDATE_APP_ID,
+  WORKSPACE_LIST_APP_ID,
 } from '../common/constants';
 import { getWorkspaceIdFromUrl } from '../../../core/public/utils';
 import { Services } from './types';
@@ -140,6 +141,17 @@ export class WorkspacePlugin implements Plugin<{}, {}, {}> {
         return null;
       }
       return React.createElement(WorkspaceMenu, { coreStart: this.coreStart });
+    });
+
+    // workspace list
+    core.application.register({
+      id: WORKSPACE_LIST_APP_ID,
+      title: '',
+      navLinkStatus: AppNavLinkStatus.hidden,
+      async mount(params: AppMountParameters) {
+        const { renderListApp } = await import('./application');
+        return mountWorkspaceApp(params, renderListApp);
+      },
     });
 
     return {};